dev_requirements.txt
coverage==7.11.0
iniconfig==2.3.0
mako==1.3.10
markdown==3.9
markupsafe==3.0.3
packaging==25.0
pdoc3==0.11.6
pluggy==1.6.0
pygments==2.19.2
pytest==8.4.2
ruff==0.14.3
pyyaml==6.0.3
